feat: implement end-to-end encryption for file uploads
